refactor: Index用Modelをmodels/indexesに集める
