+++ src/models/mod.rs
@@ -1,2 +1,3 @@
 pub mod forms;
+pub mod indexes;
 pub mod tables;
